# app/openapi.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "NEXIS Docs",
      "version": "1.0.0",
      "description": "API documentation for NEXIS"
    },
    "servers": [
      {
        "url": "https://localhost:443",
        "description": "Development server"
      }
    ],
    "paths": {
...
      "/users/login": {
        "post": {
          "summary": "User Login",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginRequest"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "User logged in successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/LoginResponse"
                  }
                }
              }
            },
            "404": {
              "description": "User not found or invalid password",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            },
            "500": {
              "description": "Unknown error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            }
          }
        }
      },
      "/users/role-login": {
        "post": {
          "summary": "User Login Role Selection",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleLoginRequest"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Role selected successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/LoginResponse"
                  }
                }
              }
            },
            "400": {
              "description": "User lacks selected role",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            },
            "500": {
              "description": "Unknown error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "LoginRequest": {
          "type": "object",
          "properties": {
            "identifier": {
              "type": "string"
            },
            "password": {
              "type": "string"
            },
            "rememberMe": {
              "type": "boolean"
            }
          },
          "required": [
            "email",
            "password"
          ]
        },
        "LoginResponse": {
          "type": "object",
          "properties": {
            "email": {
              "type": "string"
            },
            "name": {
              "type": "string"
            },
            "role": {
              "type": "string"
            },
            "availableRoles": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "token": {
              "type": "string"
            }
          }
        },
        "RoleLoginRequest": {
          "type": "object",
          "properties": {
            "role": {
              "type": "string"
            },
            "token": {
              "type": "string"
            }
          },
          "required": [
            "role",
            "token"
          ]
        },
...
        "ErrorResponse": {
          "type": "object",
          "properties": {
            "error": {
              "type": "string"
            }
          }
        }
      },
      "securitySchemes": {
        "session": {
          "type": "apiKey",
          "in": "cookie",
          "name": "session"
        }
      }
    }
  }